feat(package-managers): Add beginnings of a Gradle inspector
Start a novel approach based on a stand-alone Gradle plugin (written in
Kotlin) to analyze Gradle projects. The goal is to address shortcomings of
the current implementation:

- The `init.gradle` script cannot be debugged (only `buildSrc` and
  stand-alone plugins can [1]).

- The dynamically typed Groovy code is hard to maintain.

- Analysis of modern Android apps throws exceptions due to ambiguous
  configuration / variant selection.

- Binary artifacts are unnecessarily resolved.

- Maven is used to resolve artifacts, imitating Gradle's resolution logic.

At this stage the analysis only builds up the dependency tree to limit
the complexity of this initial implementation. Parsing of package
metadata will be added in follow-up changes.

[1]: https://docs.gradle.org/current/userguide/troubleshooting.html#sec:troubleshooting_build_logic

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>