feat(gradle-inspector): Parse package metadata from Maven POMs
Build the effective POMs for dependencies in order to parse package
metadata from them. The POMs are retrieved from the Gradle cache, so no
download via Maven is involved.

Note that the code in `MavenModelUtils` is mostly based on code from
`MavenSupport`, adjusted to work with slightly different Maven classes.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>