feat(model): Require size when writing to `ProvenanceFileStorage`
Add a `size` argument to the `putData` function of the
`ProvenanceFileStorage` which represents the size of the input stream
to be written.

This is currently a requirement to implement the `ProvenanceFileStorage`
for the ORT Server. The ORT Server uses the PostgreSQL large object
storage [1] to store binary data and the integration of that feature
requires that the size of the input stream is known. It will be
investigated if this is a hard technical requirement, and if not, the
introduced `size` argument can be removed again.

The current implementations of `ProvenanceFileStorage` do not require
the size of the input stream, but the size is easily available for all
callers of the `putData` function.

[1]: https://www.postgresql.org/docs/current/lo-interfaces.html

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>