refactor(github): Run functional tests against the published Docker image
Instead of using Batect with the committed `Docherfile`, use a GitHub
action container based on the most recent Docker image published from the
`main` branch to run functional tests. This should speed up the test run
as the Docker image does not need to be built on every run. The downside
is that `Dockerfile` changes and e.g. package manager changes depending on
it cannot be performed in the same commit / PR as the image first needs to
get published.

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>